feat: protected apis

+++ src/features/ticket/actions/delete-ticket.ts
@@ -3,12 +3,20 @@
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
 import { setCookieByKey } from '@/actions/cookies'
-import { fromErrorToActionState } from '@/components/form/utils/to-action-state'
+import { fromErrorToActionState, toActionState } from '@/components/form/utils/to-action-state'
+import { checkTicketOwner } from '@/features/auth/utils/checkTicketOwner'
 import { prisma } from '@/lib/prisma'
 import { ticketsPath } from '@/paths'
 
 export const deleteTicket = async (id: string) => {
   try {
+    
+    const isTicketOwner = await checkTicketOwner(id)
+    
+    if (!isTicketOwner) {
+      return toActionState('ERROR', 'Not authorized')
+    }
+        
     await prisma.ticket.delete({
       where: {
         id,
